相撲ゲーム0629-1
敵の移動完成

+++ SumoWrestler/Src/Application/Object/SkySphere/SkySphere.h
@@ -13,5 +13,5 @@ public:
 	void Init()			override;
 
 private:
-	int m_cnt = 0;
+	float m_cnt = 0.0f;
 };
